kconfig: Add greater/less than operations for version strings
This change to Kconfig allows two version strings to be compared,
assumedly a version token and a string literal.  This change has
no hope and no reason of ever being upstreamed since the kernel
and most other software that uses Kconfig doesn't care about
version comparisons.  This patch could be used to implement
actual arithmetic comparisons, which does have interest to some
groups, according to the threads I stumbled across when trying
to find out if anyone else already did this.

Comparison is done one subversion token at a time, comparing each
token as a number -- not as a string (i.e. "10.0" > "2.0" holds
true).  Iff both tokens match, then the next subversion token will
be considered from each string, and so on until either one token
does not match or one string runs out of tokens.

If one version string has more subversion tokens than the other,
it is compared to be greater (i.e. "3.0.0" > "3.0").  This also
means that appended versions such as -rc or -alpha make a version
considered to be greater even though this is clearly not the
logical case.

Versions which are git sha1s (in addition to being a terrible idea
for every other reason) will be considered a single subversion
token and converted to a number under the rules of atoi (the
longest uninterrupted sequence of decimal digits will be taken as
the numeric token).  Version strings which are dates will function
properly as long as the dates are delimited with either a period,
an underscore, a hyphen, or nothing at all.  Dates also must
obviously be ordered YYYYMMDD for any comparison to work properly.